feat: tiered risk-fraction and fractional-share position sizing
Adds low_capital_risk_fraction (3%, <$5,000 equity) alongside the
existing standard_risk_fraction (1%, >=$5,000), re-evaluated fresh each
call against live account equity. shares_to_buy now returns a fractional
quantity (rounded to 4 decimals, floored to 0 below $1 notional) instead
of flooring to whole shares, to support Alpaca fractional orders on a
small account. Reconciles with the small-account value cap merged just
before this: both mechanisms now operate in fractional terms.

Fixes two latent whole-share assumptions found along the way:
state_store.py crashed loading a fractional shares CSV value
(int(row["shares"])), and pipeline.py's analyst-consensus multiplier
step silently collapsed fractional results back to an integer
(round(x) with no ndigits).

+++ graywind_strategy/state_store.py
@@ -37,7 +37,7 @@ def load_state(state_dir=DEFAULT_STATE_DIR):
             for row in csv.DictReader(f):
                 state["open_positions"][row["symbol"]] = {
                     "entry_price": float(row["entry_price"]),
-                    "shares": int(row["shares"]),
+                    "shares": float(row["shares"]),
                     "stop": float(row["stop"]),
                     "target": float(row["target"]),
                     "opened_date": row["opened_date"],
